image_enhance_insert

+++ tracking_method/processing.py
@@ -131,7 +131,7 @@ def enhance_img(img):
     print("dim",img.ndim)
     print("frame shape:", img.shape, "dtype:", img.dtype)
     print("frame channels:", 1 if len(img.shape)==2 else img.shape[2])
-    img = img.copy()
+    img = img.copy().astype(np.uint8)
     #1. 2.해상도 키우기 (3~4배 정도)
     img = cv2.resize(img, None, fx=5, fy=5, interpolation=cv2.INTER_CUBIC)
     # 3 그레이스케일 변환
@@ -140,7 +140,7 @@ def enhance_img(img):
         gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
     elif len(img.shape) == 3 and img.shape[2] == 4:
         gray = cv2.cvtColor(img, cv2.COLOR_BGRA2GRAY)
-    else:
+    else:   
         gray = img 
     
     # gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
